Fix 2006 B3
The hypotheses `hSABpart` and `hLS` jointly implied `False` for slightly
subtle reasons. To be precise, because `A ≠ B` is not demanded in
`hSABpart`, `Set.encard {AB | SABpart ∅ AB}.encard` is infinite and thus
cannot equal `LS ∅` since this is finite. We thus fix the statement by
adding `A ≠ B`.

We also include some stylistic changes, specifically:
 * We express the goal using `IsGreatest`
 * We rename `SABpart` to `IsLinearPartition` since this is the
   terminology of the question statement.
 * We drop some unnecessary type ascriptions.