Fix 2011 A4
The condition `hnmat` did not refer to `nmat` at all and so the
proposition `nmat` was free (and thus the statement could be disproved).
To fix this, `nmat` has been turned into a predicate which is
characterised by `hnmat`.

In addition some minor stylistic changes are made, notably the use of
`Pairwise`.