Fix 1990 B5
The hypothesis `hanpoly` implied `False` because it did not require the
input sequence `a` to take non-zero values but it declared the resulting
polynomial always had degree `n` for each natural `n`.

I believe it would be sufficient to add the clause `(∀ i, a i ≠ 0) →` to
`hanpoly` but I believe the rewrite proposed is slightly preferable.