Fix 1989 A1
The `hpdigalt` implied false because the `=` operator was binding before
the `∧` and so it was equivalent to:
```
∀ pdig : List ℕ, (pdigalt pdig = Odd pdig.length) ∧
  (∀ i : Fin pdig.length, pdig.get i = if Even (i : ℕ) then 1 else 0)
```
meaning that it stated _any_ list alternated `1, 0, 1, 0, ...` (not just
those satisfying the `pdigalt` predicate).

This could be resolved by adding parentheses but it is slightly more
idiomatic to use `↔` for propositions (and this does have the right
precedence relative to `∧`).

Additionally the type of `putnam_1989_a1_solution` should be `ℕ∞` rather
than `ℕ` since the question does not reveal whether the set of such
primes is finite.

Some other minor style tweaks also included.