Fix 1991 A6
Before these changes `hagt` implies `False` because `=` binds before `∧` and
so `(hagt a).2` is asserted for any `a`. This could be fixed by adding
parentheses but it is more idiomatic to use `↔` rather than `=` for
propositions and this has the right precedence relative to `∧`. We make
this substitution of `↔` also in `hbg1` and `hbg2` for consistency.

Additionally, even after the fix above, `hagt` prevents length-1
partitions because in this case it reads `a 0 > a 0`. We fix this by
adding a guard for the length.

In fact this question would probably be much simpler if written using
the language of `List`s (and possibly `Nat.Partition`) but I have
avoided a complete rewrite to ease the review burden.