Fix 1967 B1
Without these changes `horder` implies `False` because it allows `i = j` and
so for example says that `L 0` is in the interior of the convex hull.

A better way to state that the points are the ordered vertices of a
convex polygon would be to replace `horder` with something like:
`∀ i j k, i < j → j < k → (∡ (L i) (L j) (L k)).sign = 1)`

However there are some missing instances that make it difficult to use
angles here.