Fix 2015 B6
Mathlib's `tsum` (with notation `∑'`) gives the wrong answer for
conditionally-convergent sums (unless they happen to sum to zero) and so
is not the right tool in this question.

This is an unfortunate footgun and gap in Mathlib.

The fix is to make a basic limit statement about the sequence of partial
sums.