Fix 2003 B3
The old statement was disprovable because it did not sufficiently
characterise the `lcmicc` function. For example the zero function
satisfies the predicate `hlcmicc`.

Mathlib should really contain an LCM function for finite sets but since
it does not exist, we have inlined a version of one in a rewrite here.

We have also dropped the requirement that `n` be strictly positive since
the theorem is in fact still true for `n = 0`.