Fix 1998 A4
The was disprovable because `Nat.digits 10 0` is `[]` rather than `[0]`.
In fact there is no reason to go back and forth with digits in the
definition of `A` so we rewrite to avoid this. We also avoid Nat
subtraction by expressing the recurrence using `n + 2` instead of `n`.

We note that the `List.reverse` is optional since 11 divides a number
iff it divides its base-10 reverse. However it seems more faithful to
the question not to embed this fact in the statement.