Fix 1996 B5
This was disprovable because the solution for `n` odd had a typo. The
minimal fix would be to fix this by changing it to:
```
(fun n : ℕ => if Even n then (3 * 2 ^ (n / 2) - 2) else (2 ^ ((n + 3) / 2) - 2))
```
but I think it is better fold the casing into the floor function as in
the official solution (to minimise chances of typos like this one). Thus
I have rewritten the solution to be identical to that provided in the
official solution.

The changes to the statement are mostly style. The only slightly
non-trivial change is to use `ℤˣ` instead of `Fin 2`.