Fix 1996 A3
This was disprovable because `studentchoicesprop` was an autoParam and
thus could be filled in with and thus could be anything. In other words,
its specification was just a default provided to downstream users who use
the lemma without specifying it, but the lemma did not actually have
this specification.

I have opted for a rewrite as there was quite a bit of room to simplify.