Fix 1995 B4
This was disprovable because `contfrac` was only defined as a root of a
quadratic equation, so we need to exclude the other root.

A really, really strict interpretation of this question might instead
define the sequence of values obtained by truncating the fraction
(`x 0 = 2207`, `x (n + 1) = 2007 - 1 / (x n)`) and a hypothesis that
this converged to a limit, and then have the goal make a statement about
this limit.

Yet another approach would be to use Mathlib's theory of continued
fractions.