Fix 2019 A4
This was accidentally trivial because the key condition was integrating
a function over a 2-dimensional surface using a 3-dimensional measure
(which is always zero) and thus was an empty condition.

Unfortunately we encounter a gap in Mathlib here (no Riemannian
manifolds) and so we are forced to use the Hausdorff measure instead of
what one would usually see (the measure on the sphere obtained from a
volume form it has as a Riemannian submanifold of Euclidean space).