Fix 1991 A4
This was disprovable because the hypothesis `hclimit` contained an
error. Specifically it should have required that infinitely-many members
of the sequence of centres meet the disc about `p` of radius `ε`, rather
than just one. As a result, `climit` was always false since we can
prove:
```
have (c) : ¬ climit c := by
  rw [hclimit, not_not]
  exact ⟨c 0, fun ε hε ↦ ⟨0, by simpa⟩⟩
```
In fact Mathlib contains a theory of limit points so we do not need
to use such a low-level description.

We provide a rewrite using a few more of Mathlib's constructions.

I note that I have embedded the fact that the area of a disc is
proportional to its radius squared in the question statement. In an
extremely strict interpretation one might demand that we merely make a
statement about the measure-theoretic `volume` of the discs.