Fix 1990 A6
This was disprovable because the `s + 1` and `t + 1` were incrementing
as elements of `Fin 10` (which carries modular arithmetic). Casting
these expressions to natural numbers should be a sufficient fix but I
have opted to provide a statement with explicit pairs (rather than a
family over `Fin 2`).

It also so happens that the new statement is computable (though out of
range of `by decide`) and can be verified to give the right answer with
`#eval`.