Fix 1989 B3
This was disprovable because `μ` was an autoParam and thus could be
anything. In other words, the intended specification of `μ` was just a
default available to downstream users of the lemma but the proof
required a proof for any `μ` (which is course is not true). We thus
promote `μ` and its specification to ordinary parameters. Other changes
are just minor style tweaks.