Fix 1982 B4
This was disprovable because the empty set satisfied the divisibility
condition. The fix is to add the `n.Nonempty` clause. Other changes are
just style.

(One might also wonder whether one should require all elements of `n` to
be non-zero since we do not usually allow division by zero but Mathlib's
conventions means this is not required.)