Fix 1968 A3
This was disprovable because the argument `ha : SDiff (Finset a)`
allowed the user to fill in any function for the set difference
function. So for example by providing a dummy set difference function
that just always returned the empty set, the goal was false.

Presumably this error arose because the original formaliser found that
they could not use the existing set difference function out of the box.
The right fix is either to put `open Classical` before the proof or to
add the typeclass argument `[DecidableEq a]`.

In the end I have provided rewrite since the original formalisation is
a bit more complicated than it needs to be.