Fix 1965 A6
This was disprovable because it did not restrict `x` and `y` to the
positive quadrant. The informal statement is quite sloppy in this same
sense (presumably because non-integral powers of negative reals are
not defined).

I have also proposed changing the implementation of "tangent to the
curve" since a priori a line could intersect a degree `m` curve in up
to `m` points so the naive set-counting definition is arguably too
strong.