Fix 1962 A2
This was disprovable because the zero function was excluded from the
solution set. The fix is to replace `a > 0` with `a ≥ 0` in the
solution. Other changes are just style.

There is incidentally a sharper version of this question which requires
establishing which solutions correspond to functions with finite /
infinite domain. However this is not actually asked.