Fix 1963 A3
This was disprovable because the `ContDiff ℝ n y` assertion in the goal
is too strong: we need to restrict to inputs `x ≥ 1` because this is the
region where we are told the behaviour of `f` is regular and even if
`f` is regular everywhere, we can see from the solution that there might
still be a singularity at `x = 0`. Furthermore, the integral was not
expressed using `intervalIntegral` and so had a kink at `x = 1` (the
point is that `∫ t in Set.Ioo 1 x, ...` will vanish for `x ≤ 1` because
`Set.Ioo 1 x` is empty).