Fix 1962 A2 (again)
This was disprovable because the condition in `P_def` does not specify
values of `f` oustide `s`. For example without this change the function
that takes the value `-1` at `0` and vanishes everywhere else would
satisfy the conditions in the goal but is not a valid solution. In fact
adding `0 ≤ f 0` should suffice but the proposed fix is preferable so as
not to rely on the "junk values" of `Real.sqrt` for negative values.

This was previously "fixed" in c10941934cae5a351b21c0f83c9709561d81749d
but unfortunately an error remained.