Fix 2022 A1
The informal solution provided at:
  https://kskedlaya.org/putnam-archive/2022s.pdf
contains some typos. The condition:
$$
  b < \ln(1 - r_{-})^2 - |a|r_{-}
$$
should be:
$$
  b < \ln(1 + r_{-}^2) - ar_{-}
$$
Similarly for the $r_{+}$ case.

The argument given in the PDF correctly derives that for $0 < a < 1$ one
should have:
$$
  b < g(r_{-})
$$
where $g(x) = \ln(1 + x ^ 2) - ax$ (and similarly for the $r_{+}$ case.
However there are some typos when transcribing this to the headline solution
(given higher up).