Fix 2020 A5
The problem was that `(a * b : Fin 3 → ℝ)` first coerces `a` and `b`
to vectors and then uses the multiplication on vectors (which is
pointwise multiplication). Thus the group structure declared in `hGgrp`
was completely ignored. This could have been fixed by writing
`(↑(a * b) : Fin 3 → ℝ)` but I think it is much clearer to use an
embedding and not depend upon arcane details Lean's coercion
elaboration.