Fix 1994 B3
This is a slightly subjective fix. The issue is that `f > 0` does not
mean that `f` is positive everywhere; it means that `f` is non-negative
everywhere and positive at at least one point. I believe this is what is
intended in the informal statement.

The issue is slightly confounded by the fact that if for all `x`,
`deriv f x > f x` and `f x` is non-negative then the derivative is
strictly positive, and thus `f` is strictly monotone, and thus being
non-negative implies being strictly positive everywhere.

So there is no mathematical change to the meaning here but I think this
is a more faithful representation of the question. It should also be
noted that the fact that there is no mathematical change depends upon
not-quite-totally-trivial facts.

I have also taken the opportunity to rewrite question with slightly
improved style.