Fix 2009 B4
This was disprovable because the integral was using the two-dimensional
measure from the plane in the integral on the one-dimensional set that
is the circle (and thus was always zero). We could fix this by supplying
the one-dimensional Hausdorff measure but this is too heavy-weight
given that we can just parameterise and use the interval integral.

A second error was the inclusion of the typeclasses `[AddCommGroup V]
[Module ℝ V]`. The subset already has this algebraic structure and by
supplying these there is ambiguity of which structure applies, as well
as the fact that these are arbitrary typeclasses with no relation to the
ambient structure on `MvPolynomial (Fin 2) ℝ`.