Fix 1963 A3 (again)
This was disprovable because `y` could fail to be differentiable at `1`.
E.g., one could define `y` such that:
`y x = ∫ t in (1 : ℝ)..x, putnam_1963_a3_solution f n x t` for `x ≥ 1`
and `y x = 0` for `x < 1`. This will satisfy the RHS of the iff in the
goal but not the left.

I see two obvious ways to resolve the issue:
 1. Demand that `y` is C^n on `[1, ∞)`
 2. Require that `y x = ∫ t in (1 : ℝ)..x, putnam_1963_a3_solution f n x t`
    holds on a neighbourhood of `[1, ∞)`

The informal statement is very vague so I have somewhat arbitrarily
opted for the first of these resolutions.

This was previously "fixed" in 7a9215271b97b0d3a87f747c3828fea29e0d9ef7 but unfortunately an error remained.