Fix 1984 A6 (again)
This was disprovable because the value `g 0` breaks the period-4
pattern. More precisely, the period 4 arises because powers of 2
mod 10 have period 4 (2, 4, 8, 6, 2, 4, 8, 6, ...). However this
starts at 2^1 and the inital value 2^0 = 1 breaks the pattern.

This was previously "fixed" in 295b586c8791871fd1dcf1a83742e4d2b56fb499
but unfortunately an error remained.