Fix 2023 B3
This was disprovable because the `IsZigZag_def` predicate was missing
parentheses. Because of this, the following was in fact true:
```lean
have (k : ℕ) [NeZero k] (y : Fin k → ℝ) : k = 1 ↔ IsZigZag y := sorry
```
The issue being that for all sufficiently large `i`, `i + 2 < k` was
`False` and thus `i + 2 < k → y (i + 2) < y (i + 1)` was `True` and
thus within the second clause of the disjunction in `IsZigZag_def`, we
have `y i < y (i + 1)` for all sufficiently large `i : ℕ`. However `y`
is a finite sequence and so cannot decrease forever. Thus the second
clause was always `False` so that the predicate collapsed to just
`k = 1`.