Fix 1989 B2
This had contradictory hypotheses and had a trivially provable goal.

The hypotheses were contradictory because the `pow` function was
universally defined but required to satisfy axioms for each semigroup
structure that a type might carry. Thus if we invoke the hypotheses
on a type `S` twice, once for one semigroup structure and once for
another, we obtain a contradiction unless they happen to agree about
`x ^ n` for all `x` and `n`.

The goal was trivial because it amounted to a statement of the form
`... → Nonempty (Group S)` which is trivial if `S` is non-empty.